Added data averaging to the output diagnostics.
These may be specified by using the 'average' keyword when
specifying a dump variable. The non-averaged variable will still
be written to restart dumps but not full or normal dumps. If you
also want the non-averaged variable to be written then you can
add the 'snapshot' option.
The averaging window can be specified using the
'averaging_period' keyword. If you require a minimum number of
cycles over which to perform the averaging, you can give the
'min_cycles_per_average' which will adjust the timestep to ensure
that the condition is met.

begin:output

  dt_snapshot=1.0e-15
  full_dump_every=10
  restart_dump_every=100
  force_final_to_be_restartable=T
  averaging_period=1.0e-17

  px=always + average + snapshot
  py=full + average + snapshot
  pz=full + average

end:output

With this configuration, px will be written in both normal and
averaged form at normal, full and restart dumps. py will be
written in both forms at full and restart dumps. Since pz is not
required for restart, only the average will be written at both
full and restart dumps.