A better fix for autoloading a discontinuous density profile.
There are difficulties in smoothly assigning particle weights when
using per-particle weights and the density profile is discontinuous.
The previous fix for this issue was to load particles in the cells
directly adjacent to the discontinuity. The new fix is to reflect
the weight assignment in vacuum cells back into the non-vacuum area.

Also improved the efficiency of randomly loading particles.