Adjusted polarization angle and defaults for lasers.
The polarization angle for a laser is now measured in a systematic
way with respect to the right-hand triad of propagation direction,
electric and magnetic fields. The previous version was somewhat random.
If the laser is on x_min then the default E field is in the y-direction
and the B field is the z-direction. The polarization angle is measured
clockwise about the x-axis with zero along the y-axis.
Similarly, for propagation directions:
 y_min: angle about y-axis, zero along z-axis
 z_min: angle about z-axis, zero along x-axis
 x_max: angle anti-clockwise about x-axis, zero along y-axis
 y_max: angle anti-clockwise about y-axis, zero along z-axis
 z_max: angle anti-clockwise about z-axis, zero along x-axis

Also set new default values for lasers. The default end time is t_end
and the default profile is 1.0.