Added "nstep_snapshot" entry to the output block of input.decks
This allows a user to specify the number of timesteps between
output dumps as well as the simulation time between dumps.
Both can be specified and they will both be tested for.
Also made "dt_snapshot" an optional parameter.
If "dt_snapshot" is not specified then it is not tested for.
If "nstep_snapshot" is not specified then it is not tested for.