Made particle probe parameters consistent.
Previous versions had differing ways of specifying probes in
the input deck for each of epoch{1,2,3}d. They also had different
implementations in the code for testing if a particle had crossed
a probe.
This patch makes them consistent. They are all specified using
a point in the plane and the normal vector to the plane. Particles
are recorded if they cross the plane in the direction of the normal
vector.
The points and normals are given using the following syntax:
  point = (1,2.5,3)
  normal = (0.5,1,0)

In 1d, both the above and scalar form are accepted:
  point = 2.5
  normal = 1