Simplified deck parsing process.
The input deck is now only parsed twice. The first time
it is called with deck_state set to "c_ds_first" and this pass
deals with anything that does not require pre-allocated memory.
On the second pass, deck_state is set to "c_ds_last" and this
deals with entries that could not be parsed in the first sweep.