Output a list of each type of output dump for VisIt.
The VisIt visualisation tool will accept a list of files
which belong together and make these available for plotting.
Since "normal", "full" and "restart" dumps each contain different
data, it is useful to have a list for each type.
This may also be useful for the other plotting libraries in future.