Added "PLAIN_DERIVED" and "POINT_DERIVED" block types.
These block types are intended for use by the VisIt reader
plugin and should never be found in an actual SDF file.
They are to be used when a derived variable can be generated
based on variables found in the SDF file.
They are similar to the "STITCHED" block types except that
they contain a callback function "populate_data" which is called
whenever the data block needs to be populated.
The "variable_id" entries are used to indicate any blocks which
should be filled before calling the callback function.