Enable multiple "output" blocks in the input deck.
It is now possible to have multiple output blocks, each
with their own "dt_snapshot" or "nstep_snapshot" and their
own set of output variables.
The syntax remains the same as before with the addition of
"name" and "restartable" fields.
The "name" field specifies the file name to use for the output
list. Each time EPOCH generates an output dump, it writes an
entry into the file "<name>.visit". This can be used to find
all the output dumps of a specific output block.
The "restartable" field specifies that the output block should
generate ouput dumps containing all the information necessary
to restart a simulation.