Implemented the "ejected_particles" output parameter.
If the "ejected_particles" variable is requested in the
output block then all the particles which have left the
simulation domain are written to file.
The parameter is assigned a dumpmask in the same way as
other variables in the output block.
Once the data has been written, the ejected particle lists
are reset and will accumulate particles until the next
requested output dump.