Fix for field boundary conditions on moving window.
The open boundary conditions do not set values for ghost
cells at the right edge of the domain. This becomes a problem
for moving windows since the ghost cell values are shifted into
the simulation domain. This patch fixes the problem by
interpolating values into the ghost cell before shifting the
grid.