Added single and double precision routines to SDF.
The SDF library now contains both single and double precision
versions of every subroutine. This allows the library to be
completely divorced from the EPOCH code base as it no longer
needs to be compiled with the same precision for real numbers.
The precision of the data to be written is determined by the
data type passed in the subroutine arguments.

This change also provides the necessary infrastructure for
writing single precision output from a double precision run.