Added global particle IDs.
Particle IDs are useful if you want to track the progress
of each particle throughout the simulation. Since they increase
the size of each particle data structure, they are disabled by
default and must be enabled using a compiler flag.
The PARTICLE_ID flag will use an 8-byte integer to represent the
ID and PARTICLE_ID4 uses a 4-byte integer.
They are written to file using the "id" flag.
Eg.
    particles = always
    id = always

in the "output" block will write particle IDs to file.

Note: In the current implementation, the particle IDs are
passed between processors and written to file using REAL numbers.
This means that in double precision the maximum particle ID
is 2^53 ~ 10^16. This should be ample for the forseeable future.
However, if the code is compiled for single precision then the
maximum ID is 2^24 = 16777216. Probably not big enough.