Added "npart_per_cell" to species blocks.
This patch adds the ability to specify the number of particles
per cell to use for the initial particle loading.
In the current implementation, "npart_per_cell" is a simple
integer constant. At a later stage this will be extended to
allow "npart_per_cell" to be a spatially varying function.

If per-species weighting is used then the value of
"npart_per_cell" will be the average number of particles per
cell.  If "npart" or "frac" have also been specified for a species,
then they will be ignored.

To avoid confusion, there is no globally used "npart_per_species".
If you want to have a single value to change in the input deck
then this can be achieved using a constant block.