Allow missing particle species in restart dumps.
When restarting from a dump file, EPOCH scans the particle
species in the file and tries to match them up with species
in the input deck. There are legitimate reasons why they
wont match up. For example, ejected particles appear as a
new species in the dump files but are never listed in an
input deck.
This patch allows such anomolies to occur and just issues
a warning.