Always allocate the number of particles specified.
Before this patch, the code would round to the nearest number
of particles on each process. In cases where there was less than
one particle per process, no particles would be assigned.
In the case of tracer particles and a large number of processors,
it is not unreasonable to want less than one particle per
processor.