Truncate strings which are too long for SDF.
Some of the output block names are based on species names
provided by the user which may exceed the maximum string
length allowed by the SDF library.
In the case of ID fields, the string is silently truncated
and the SDF library mangles the name to ensure it is unique.
For name fields a warning message is issued.