Added single-precision output for EPOCH.
This adds a new 'single' field to the dumpmask, allowing
output variables to be converted to single precision when being
dumped to file. This is specified on a per-variable basis.

eg.
begin:output
  dt_snapshot = 8 * femto

  grid = always
  ex = always
  ey = always + single
end:output

In this example, the grid variable 'ex' will be written as a
double precision array and 'ey' will be converted to single
precision.