Changes to the QED and collision routines.
Collision routines are now capable of dealing with cold plasmas.
Added new parameter photon_dynamics to QED routines. This
parameter means that the output is essentially a diagnostic of
photon production while reducing runtime.
Some of the constants used in QED routines are replaced by
numerical values to allow the code to run in single precision.