Changes and fixes to the parallel SDF library.
The SDF C-library now performs parallel factorisation for a
data block when the block info is read. This ensures that the
correct dimensions are present when it comes to allocating
storage for the block.
Also, the conversion from double to single precision has been
moved into the sdf_helper_read_array routines so that extra
storage only needs to be allocated if conversion occurs.