Added the "use_exact_restart" control block parameter.
This logical flag may be set to "T" or "F". By default it is "F".
If this flag is set in conjunction with "restart_snapshot", EPOCH
will attempt to restart the code using the exact configuration it
was using when the restart dump was produced.
The domain split amongst processors will be identical, as will
the particle lists. The code will also reset the status of the
random number generator on each process.
Note that the flag will be ignored if the number of processors
does not match that used in the original run.