Add a routine for getting truncated SDF IDs.
The sdf_get_block_id(h,long_id,short_id) function looks for
a block with the "long_id" untruncated SDF ID string.
If found, the shortened version of the string is copied into
"short_id" and the function returns ".TRUE.".
If not found, "long_id" is truncated and copied into "short_id"
and the function returns ".FALSE.".

ID strings which have been truncated are often mangled so that
they remain unique within the file. Therefore it may not be
possible to locate the block using just a truncated ID string.