Use an untruncated ID string to find point meshes.
When writing point variables, we also pass through the ID string
of the associated point mesh. If this was too long and has been
truncated, the shortened version may no longer uniquely identify
the correct mesh block. To avoid this problem, we use the
untruncated version of the string to request the shortened version
from the SDF library