Ionization species inherit "dumpmask".
The species which are automatically created for each ionization
level are now given the same "dumpmask" or "dump" parameter as
the parent species. This behaviour can be overridden by explicitly
adding a species block of the same name with a differing dumpmask.
eg.

begin:species
  name = Helium
  ionisation_energies = (24.6*ev,54.4*ev)
  dump = F
end:species

begin:species
  name = Helium1
  dump = T
end:species