Rationalised the usage of multiple "output" blocks.
This set of changes introduces a clear distinction between the
old style of output block and the new style.

With the old method, there is a single unnamed output block. It
has three levels of dump frequency: "normal", "full" and
"restart". There is a single dump interval specified by either
"dt_snapshot" or "nstep_snapshot" and the frequency between full
and restart dumps is specified in relation to this interval using
"full_dump_every" and "restart_dump_every".

With the new method, there can be multiple output blocks and each
one has a "name" field to distinguish it. It is the existence or
absence of this field which determines whether it is a new or old
style output block. Each output block has a single dump frequency.
In order to have multiple dump levels you must create a separate
output block for each level. Since there is no longer the notion
of multiple dump levels, the "full_dump_every", etc. flags are
not supported.
There are a few output parameters which only make sense globally.
When using the new style of output block, these are now specified
by placing them in their own "output_global" block.
The list of fields which are now specified globally are:
  force_first_to_be_restartable
  force_final_to_be_restartable
  use_offset_grid