Reverted "Fix for per-species averaging."
This reverts the patch made on Wed Oct 24 14:19:41 2012. I'm
not sure exactly what bug that change was intended to fix, but
it breaks multiple output block dumping.
Specifically, whenever a dump is triggered, the iomask is set
to be the combination of dumpmasks for all output blocks. This is
incorrect behaviour. The correct behaviour is that the iomask is set
to be the combination of dumpmasks for all output blocks which have
triggered a dump to occur.

Once I have found the "per-species averaging" bug, I will come up
with a new fix for it.