Added "dump_cycle" parameter to the "output" block.
This adds the "dump_cycle" integer parameter to the "output"
block in the input deck.
If set to a positive integer then the output file number will
be reset to zero after the specified cycle number is reached.
The default is "0", so dump cycling never occurs.
eg.
block:output
  nstep_snapshot = 10
  dump_cycle = 2
  ex = always
end:output

With this block, output will be written at step 10 to the output
file "0000.sdf", step 20 to "0001.sdf", step 30 to "0002.sdf",
step 40 to "0000.sdf", step 50 to "0001.sdf", etc.