Allow "restart_snapshot" to accept filenames.
The "restart_snapshot" parameter in the "control" block of the
input deck will now accept a filename instead of a cycle
number.
If you want to restart from "0012.sdf" then it can either be
specified using:
  restart_snapshot = 12

as in previous versions, or alternatively it can be specified
using:
  restart_snapshot = 0012.sdf

This syntax will be useful once output file prefixes have been
implemented.