Added "rolling_restart" flag to the output block.
If the "rolling_restart" flag is set to "T" in a named output
block then it automatically sets the output block to have the
properties required for performing rolling restarts.
It is currently just a shorthand for setting the following flags:
  dump_cycle = 1
  restartable = T
  file_prefix = roll

In the future, there may be some additional properties attached
to rolling restart dumps and these will also be enabled using
this flag.