Write a unique list to the "*.visit" files.
Each time an output dump is written, the name of the output
dump gets written to a "<name>.visit" file, where <name> is
the name of the corresponding output block.
When dump cycling is enabled, this could lead to multiple
repetitions of the same output files in the list.
The diagnostic routine now keeps a list of the filenames already
written and checks for duplicates before writing to the *.visit
file.