Reduce the number of CPUs if necessary.
The previous patches enforce a minimum domain size. In some
situations it may not be possible to divide the simulation
amongst all the processors and still meet this restriction.
In such circumstances, we keep reducing the number of processes
until we find a suitable split.
On a typical large unattended run this behaviour is better than
exiting and losing allocation slot.

To have the simulation exit under such circumstances, add the
flag "allow_cpu_reduce = F" to the control block.