Added 'x_grid_min', etc. for the simulation domain.
x_min, x_max, etc. now refer to the simulation domain as
specified in the input deck.
On the other hand, x_grid_min variable refers to the position of
x(1). This is not the same as x_min, since the grid is defined at
cell centres so everything is shifted by half a cell. In addition
to this, when CPML boundaries are used the whole grid is shifted
by the thickness of the CPML boundary.

Here is an ASCII representation of the grid without CPML
boundaries:
                 A                               B
-|---o---|---o---|---o---|---o---|---o---|---o---|---o---|---o---|-
     -1      0       1          ...         nx      nx+1    nx+2

Here 'o' is a cell centre and '|' is a node-centre. The position
marked (A) is x_min - ie. the x_min specified in the input deck.
The position marked (B) is x_max.
x_grid_min is the position marked by x(1). This is the location
used when calculation the cell number in which a particle is
located. x_grid_max is the position marked by x(nx). x(0) and
x(nx+1) are ghost cells.

Here is an ASCII representation of the grid with CPML
boundaries:
                   A                       B
-|--o--|--o--|--o--|--o--|--o--|--o--|--o--|--o--|--o--|--o--|-
    1     2     3     4       ...     nx+3  nx+4  nx+5   nx+6

In this diagram, the CPML boundaries have a thickness of 3.
Here, (A) and (B) are x_min and x_max as before. Now (1) is x(1)
which is x_grid_min. (4) is the first cell of the actual domain
but due to the CPML layer the grid position is x(1+3). The last
domain cell is at x(nx+3) and x_grid_max is x(nx+2*3).