Only convert point data when necessary.
There was a bug when the SDF library was instructed to convert
point arrays from double to single precision before writing to
file. The entire buffer array was converted even though it is
only necessary to convert the data being written. This led to
output dumps being much slower than they should be.