Added "filesystem" option to output blocks.
Some filesystems can be unreliable when performing parallel I/O.
Often this is fixable by prefixing the filename with 'ufs' or
'nfs'. This change adds an option to both the "output" block
and "output_global" block enabling the filesystem prefix to be
used. eg:

begin:output_global
  filesystem = ufs
end:output_global