Added particle gamma and relativistic mass output.
It is now possible to output per-particle gamma and per-particle
relativistic mass. These are specified in the same way as other
particle variables in the "output" block using the parameters
"gamma" and "relativistic_mass". eg.

begin:output
  dt_snapshot = 1e-15
  particles = always
  gamma = always
  relativistic_mass = always
end:output

There is also now a "rest_mass" parameter which is just an alias
for "mass", controlling the output of per-particle rest mass.