Allow "qed" blocks when -DPHOTONS is not used.
Previous versions of the code would halt execution if a "qed"
block was found in the input deck for code which was not compiled
using the -DPHOTONS preprocessor flag. This can make it difficult
to use the same input deck with and without QED enabled.
The code now only halts if "use_qed=T" inside the "qed" block.