Add reading of the data directory name from a file.
EPOCH now first checks for the existence of a file named
"USE_DATA_DIRECTORY" in the current working directory. If such
a file exists, it reads it to obtain the name of the data
directory to use. If no such file exists, it prompts for a
data directory name as before.
This is useful for cluster setups in which it is difficult or
impossible to pipe in the directory name using a job script.