Check for the existence of a "STOP" file.
EPOCH will now check for the existence of a file named either
"STOP" or "STOP_NODUMP" in the simulation output directory.
The check is performed at regular intervals and if such a file
is found then the code exits immediately. If "STOP" is found
then a restart dump is written before exiting. If "STOP_NODUMP"
is found then no I/O is performed.

The interval between checks is controlled by the integer parameter
"check_stop_frequency" which can be specified in the "control"
block of the input deck. If it is less than or equal to zero then
the check is never performed.
The default value is 10.