Stop execution after a given elapsed walltime.
This adds a few parameters to the "control" block of the input
deck which allow the code to halt execution and create a
restart dump after a specified walltime has elapsed.

Four parameters have been added:

"check_walltime_start" specifies a simulation time after which
to begin checking whether the walltime has been reached.

"check_walltime_frequency" specifies how frequently to check the
walltime. This is given as the number of simulation iterations
between checks.

"stop_at_walltime" is the walltime at which to halt the simulation.
This is specified as the number of seconds since the start of the
simulation.

"stop_at_walltime_file" is the filename from which to read the
value for "stop_at_walltime". Since the walltime will often be
found by querying the queuing system in a job script, it may be
more convenient to pipe this value into a text file rather than
modifying the input deck.