Added routine for automatically checking walltime.
If a positive value for "stop_at_walltime" is specified in the
control block of an input deck then the code will halt once this
time is exceeded and write a restart dump.
The previous mechanism for testing the condition involved syncing
all processes at regular intervals. With this change, each process
checks the condition independently and only attempt to synchronise
once the condition has been met.