Added extra output_point routines for backwards compatibility.
One of the previous changes added a 'species_id' field to all
sdf_output_point routines. This change adds back an interface
which does not require the field so that older codes need not
be changed.
If the old-style interface is used then 'nospec' is used for the
'species_id' field.