Added "allow_missing_restart" flag.
This adds the flag "allow_missing_restart" to the control block
of input decks. The default value is "F".
When "restart_snapshot" is specified then the simulation first
checks that the specified restart dump is valid. If the restart dump
exists and is valid then it is used to provide initial conditions
for the simulation.
However, if the restart dump does not exist or is not usable for
some reason then by default the simulation will abort.
If "allow_missing_restart" is set to "T" then the simulation will
not abort but will continue to run and use the initial conditions
contained in the input deck to initialise the simulation.