EPOCH now gets all version information from git.
There is no longer any version numbering hard coded into EPOCH.
The version number is derived from the "COMMIT" file which must
always be present at build time. In a git repository, this file
is created by querying the repostory with "git describe".
In a tarball, the file must already exist. In order to ensure that
this is the case, tarballs should be created using "make tarball"
from within a git repository.