Added data reduction for grid-based variables.
This change adds the integer parameters "skip_{x,y,z}" and "skip" to the
subset blocks. If not present the default value is 0.
If a parameter is set to a positive integer then all grid-based
variables using the subset restriction will be reduced when being written to
file. This is achieved by skipping by the specified number of cells in
each of the specified directions. The "skip" parameter provides a quick
method for setting the same number of cells to skip in all directions.

This currently only applies to grid-based variables and is ignored for
data averages.