adding an automatic example running interface 1D
every example should be in its own subfolder under examples/. It will be
run by a makefile and postprocessed by a postprocessing script inside
the same folder. This allows to easily run examples and compare their
output data. Its not really a test suite, but it makes it easy to see if
an example changes its behaviour or refuses to start due to some change
in the code.