add pre-commit
this file should be linked as the git pre-commit hook. Git will run the
script on every 'git commmit' and abort the commit process if the
exection errors. Therefore the script is only testing if the code
builds, not run any tests!
RUN THEM BY YOURSELF!
To bypass the pre-commit hook exection use 'git commit --no-verify'