Allow dump file to be requested at run time
The code now monitors for the existence of a file named "DUMP"
in the Data directory. This is polled for with the same frequency
as "STOP" and "STOP_NODUMP". If the DUMP file is found then an
output dump is written. If the file contains the name of an output
block then that will be the type of dump file written, otherwise
a restart dump is written.