Added "dump_at_walltimes" flag to the output blocks
This flag specifies a list of times at which to perform an output
dump for the current block. These times are the total elapsed time
in seconds since the start of the simulation. Note that if the
simulation has been restarted then the total elapsed time will
include the accumulated walltime of all previous runs that were
used to produce the restart dump.