Add dlb_force_interval parameter to the input deck
This adds the integer parameter "dlb_force_interval" to the control
block of the input deck. It determines the maximum number of timesteps
to allow between forcing a full load balance sweep.
If the current load balance is greater than the value of "dlb_threshold"
then the load balancer exits before attempting to improve the loading.
If "dlb_force_interval" is greater than zero, then the full load
balancer will be run at the requested interval of timesteps, regardless
of the value of "dlb_threshold". Note that the simulation will only be
redistributed if this would result in an improved load balance.
The default value of "dlb_force_interval" is 2000.