Add dlb_maximum_interval parameter to the input deck
This adds the integer parameter "dlb_maximum_interval" to the control
block of the input deck. It determines the maximum number of timesteps
to allow between load balancing checks.
Each time that the load balancing sweep is unable to improve the load
balance of the simulation, it doubles the number of steps before the
next check will occur. It will keep increasing the check interval until
it reaches the value given by "dlb_maximum_interval". If the value of
"dlb_maximum_interval" is negative then the check interval will increase
indefinitely. When the load balancing sweep finds an improvement to the
load balance of the simulation, the check interval is reset to one.
The default value of "dlb_maximum_interval" is 500.