Add balance_first parameter to the input deck
This adds the logical parameter "balance_first" to the control
block of the input deck. It determines whether a load balance
will be attempted on the first call of the load balancer.
The load balancer performs to functions: first it attempts to find a
domain decomposition that balances the load evenly amongst processors.
Next, it redistributes the domain and particles onto the new layout (if
requred). This latter step is always required when setting up the
simulation, so the load balancer is always called once during set-up.
This flag controls whether or not a load balance is attempted during
this call, regardless of the value of dlb_threshold.
The default value is TRUE.