Allow variables named e_ where _ is a number
The parser has a specific check for the letter e followed by a number to
allow for 5e6, etc. However, this rule prevented the use of any symbolic
constant named e1 or similar. This patch adds a new check to see if the
letter e was both preceded and followed by a number.