Reduce the minimum domain size
This sets the minimum domain size to be equal to the number of cells
needed for current deposition on an internal boundary.
This is half the particle shape width plus one cell due to the distance
a particle can travel in a timestep and one extra cell due to grid
staggering.
This minimum is required in order to correctly communicate ghost cell
information in a single step. The minimum domain size could be reduced
further by repeating the communications.