Automatically create pull requests for merge cascade
When a devel branch is updated then it should also get merged into the
other devel branches. ie. a 4.18-devel updated should get merged into
4.19-devel, etc.

Note that this workflow relies on a dummy user account to auto-approve
the pull request. This account is authenticated by adding an
environmental variable secret to the repository named "GH_APPROVE" which
is set to the personal access token for the dummy account.